requirements.txt
Pillow
beautifulsoup4
cfscrape
colorama
coloredlogs
lxml
requests